+++ Code/VFSPrototype/VFSBase/Persistance/BlockParser.cs
@@ -5,30 +5,27 @@ using System.Text;
 using System.Threading.Tasks;
 using VFSBase.Implementation;
 using VFSBase.Interfaces;
+using VFSBase.Persistance.Blocks;
 
 namespace VFSBase.Persistance
 {
     internal class BlockParser
     {
-        public static IIndexNode ParseBlock(byte[] bb, FileSystemOptions options)
+        private readonly FileSystemOptions _options;
+
+        public BlockParser(FileSystemOptions options)
         {
-            if (bb.Length != options.BlockSize) return new EmptyBlock();
+            _options = options;
+        }
+
+        public IIndexNode ParseBlock(byte[] bb)
+        {
+            if (bb.Length != _options.BlockSize || bb[0] == 0) return new EmptyBlock();
+
+            var type = bb[0];
+            
+
             return null;
-        }
-    }
-
-    internal class EmptyBlock : IIndexNode
-    {
-        public string Name
-        {
-            get { return ""; }
-            set { }
-        }
-
-        public Folder Parent
-        {
-            get { return null; }
-            set {  }
         }
     }
 }
