+++ Code/VFSPrototype/VFSBase/VFSBase.csproj
@@ -66,6 +66,7 @@
     <Compile Include="Implementation\FileSystemTextManipulator.cs" />
     <Compile Include="Implementation\Folder.cs" />
     <Compile Include="Persistance\BlockParser.cs" />
+    <Compile Include="Persistance\Blocks\EmptyBlock.cs" />
     <Compile Include="Persistance\PositionCalculator.cs" />
     <Compile Include="Persistance\PositionCalculatorSettings.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
